+++ R/s3.R
@@ -138,3 +138,44 @@ print.box <- function(x, ...) {
   cat_line("<box>")
   print(unbox(x))
 }
+
+
+#' Create zap objects
+#'
+#' @description
+#'
+#' `zap()` creates a sentinel object that indicates that an object
+#' should be removed. For instance, named zaps instruct [env_bind()]
+#' and [call_modify()] to remove those objects from the environment or
+#' the call.
+#'
+#' The advantage of zap objects is that they unambiguously signal the
+#' intent of removing an object. Sentinels like `NULL` or
+#' [missing_arg()] are ambiguous because they represent valid R
+#' objects.
+#'
+#' @param x An object to test.
+#'
+#' @export
+#' @examples
+#' # Create one zap object:
+#' zap()
+#'
+#' # Create a list of zaps:
+#' rep(list(zap()), 3)
+#' rep_named(c("foo", "bar"), list(zap()))
+zap <- function() {
+  `zap!`
+}
+#' @rdname zap
+#' @export
+is_zap <- function(x) {
+  inherits(x, "rlang_zap")
+}
+
+`zap!` <- structure(list(), class = "rlang_zap")
+
+#' @export
+print.rlang_zap <- function(x, ...) {
+  cat_line("<zap>")
+}